Added lsg tool, which shows git status in ls style.
We process the ls output to give a git status inline with the file
information. It's not perfect, but I'm really quite pleased with it.
We show the status of the individual files if they are listed, and
include the status of the repos, with the branch they are currently
on.

There is probably more that can be done, but this looks pretty good
to me right now.