Update lsg to allow the 'ls' used to to be changed.
The ls tool that the command uses can be replaced by a different
one - say, so that the GNU ls might be used on MacOS, for example.
At present this appears to work, although there's no configuration
option to change it; it must be changed in the code at the current
time.