Add lsg support for quoted and filetyped output.
Quoted output is able to be used on the GNU ls, and whilst not common,
actually makes the parsing a lot more reliable. The filetyped output
is more common, and makes things a little more complex. However,
stripping both off is achievable without too much difficulty, once
it's seen that those switches are present.