Slightly better support for bash 3; simplify some statements.
In bash 3, -1 cannot be used for the last character of a string
in a substring. So the code needs to be a more explicit.

Some of the statements were raising lint errors about escapes
being redundant (which is intentional, because they're being
fed into PS1), so these have been made more explicit.