Fix lsg -R failing to get the git status of non-top level.
The -R option would fail to construct the correct path when determining
the recursed files. The result would be that it failed to display any
of the git status for files not at the top level. We now reconstruct
the correct set of git parameters so that we can see the git status
of the files in the subdirectories.

This complements the --nest option quite nicely, as it doesn't indent
as much, but does give a nice lined up output for the git status. The
downside is that because we're not re-caching the status for each of
the subdirectories, we run a bit slower.