Add support for --help and --version.
Bug https://github.com/gerph/ls-with-git-status/issues/1 reported that the
help was garbled. The tool was trying to parse the output from the --help
as if it were a list of files. Silly!

This change recognises the --help and the --version and outputs the messages
from ls and then one from lsg itself to make it clear that this is a little
different. The --version also includes a SHA1 so that it is hopefully able
to be differentiated, even if it's not actually a version number.