Add environment variables LSG_OPTIONS and LSG_COMMAND.
The LSG_COMMAND environment variable allows the command run to be
overridden, if you want to use a different `ls` tool.

The LSG_OPTIONS environment variable allows the default set of
options to be overridden. This means that you can force the tool
to use long listing by default or any of the other forms.

Closes #3 from klaernie.